[Gold I] Title: K번째 수, Time: 120 ms, Memory: 14264 KB -BaekjoonHub

+++ 백준/Gold/1300. K번째 수/K번째 수.java
@@ -10,7 +10,7 @@ public class Main {
 
     System.out.println(two(n, m));
   }
-  public static long two(int n, int m) {
+  public static int two(int n, int m) {
     int left = 0;
     int right = Integer.MAX_VALUE;
